Modify IBM script and data procesing
Needed to modify a lot of the data processing steps in order to save
all the correct plot ID, plant size, and year information in the stan
models.  This is then read into the IBM script and used to generate
predictions for cover at the plot level in each year.

Additional steps are used to get the cover for the start of time
series where cover cannot be predicted but is still useful for
context.

Care needs to be taken when matching quadrats and when rescaling
predictions from the modeled scale to the measured scale.